backend/requirements.txt
# Web Framework
fastapi==0.115.0
uvicorn[standard]==0.32.0
python-multipart==0.0.12

# Database
sqlalchemy==2.0.35
psycopg2-binary==2.9.10
alembic==1.13.3

# AI & ML
openai==1.54.3
langchain==0.3.7
langchain-openai==0.2.8
langchain-community==0.3.7
tiktoken==0.8.0
pinecone-client==5.0.1
tensorflow==2.18.0
scikit-learn==1.5.2
numpy==1.26.4
pandas==2.2.3

# PDF Processing
pypdf2==3.0.1
pdfplumber==0.11.4
pytesseract==0.3.13
pillow==11.0.0

# API & Async
httpx==0.27.2
aiohttp==3.11.7
requests==2.32.3

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.2.1
python-dotenv==1.0.1
pydantic==2.9.2
pydantic-settings==2.6.1

# Monitoring & Logging
sentry-sdk==2.18.0
python-json-logger==3.1.0

# Data Validation
email-validator==2.2.0

# Financial Data
alpha-vantage==2.3.1
yfinance==0.2.48

# Task Queue
celery==5.4.0
redis==5.2.0

# Testing
pytest==8.3.3
pytest-asyncio==0.24.0
pytest-cov==6.0.0
httpx==0.27.2

# Utilities
python-dateutil==2.9.0
pytz==2024.2
jinja2==3.1.4
